+++ examples/azure-http-function-v4/HttpExample/app.js
@@ -1,0 +1,99 @@
+const path = require('path')
+const express = require('express')
+const bodyParser = require('body-parser')
+const cors = require('cors')
+const compression = require('compression')
+const { getCurrentInvoke } = require('../../../src/index') // require('@vendia/serverless-express')
+const ejs = require('ejs').__express
+const app = express()
+const router = express.Router()
+
+app.set('view engine', 'ejs')
+app.engine('.ejs', ejs)
+
+router.use(compression())
+router.use(cors())
+router.use(bodyParser.json())
+router.use(bodyParser.urlencoded({ extended: true }))
+
+// NOTE: tests can't find the views directory without this
+app.set('views', path.join(__dirname, 'views'))
+
+router.get('/api', (req, res) => {
+  const currentInvoke = getCurrentInvoke()
+  const { event = {} } = currentInvoke
+  const { requestContext = {} } = event
+  const { domainName = 'localhost:7071' } = requestContext
+  const apiUrl = `https://${domainName}`
+  res.render('index', { apiUrl })
+})
+
+router.get('/api/vendia', (req, res) => {
+  res.sendFile(path.join(__dirname, 'vendia-logo.png'))
+})
+
+router.get('/api/users', (req, res) => {
+  res.json(users)
+})
+
+router.get('/api/users/:userId', (req, res) => {
+  const user = getUser(req.params.userId)
+
+  if (!user) return res.status(404).json({})
+
+  return res.json(user)
+})
+
+router.post('/api/users', (req, res) => {
+  const user = {
+    id: ++userIdCounter,
+    name: req.body.name
+  }
+  users.push(user)
+  res.status(201).json(user)
+})
+
+router.put('/api/users/:userId', (req, res) => {
+  const user = getUser(req.params.userId)
+
+  if (!user) return res.status(404).json({})
+
+  user.name = req.body.name
+  res.json(user)
+})
+
+router.delete('/api/users/:userId', (req, res) => {
+  const userIndex = getUserIndex(req.params.userId)
+
+  if (userIndex === -1) return res.status(404).json({})
+
+  users.splice(userIndex, 1)
+  res.json(users)
+})
+
+router.get('/api/cookie', (req, res) => {
+  res.cookie('Foo', 'bar')
+  res.cookie('Fizz', 'buzz')
+  res.json({})
+})
+
+const getUser = (userId) => users.find(u => u.id === parseInt(userId))
+const getUserIndex = (userId) => users.findIndex(u => u.id === parseInt(userId))
+
+// Ephemeral in-memory data store
+const users = [{
+  id: 1,
+  name: 'Joe'
+}, {
+  id: 2,
+  name: 'Jane'
+}]
+let userIdCounter = users.length
+
+// The serverless-express library creates a server and listens on a Unix
+// Domain Socket for you, so you can remove the usual call to app.listen.
+// app.listen(3000)
+app.use('/', router)
+
+// Export your express server so you can import it in the lambda function.
+module.exports = app